main/api/serializers.py
from main.models import Notification
from rest_framework import serializers


class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notification
        fields = ['id','title', 'image', 'text', 'created','new','unread']
